Initial LCP support
This adds the initial support for LCP in that it sets up the right
repositories and imports the correct LCP dependencies. It adds code
to retrieve an LcpService, if the LCP library is available, and
adds the needed format names and flags in order to ensure that
books are seen in feeds.

Affects: https://www.notion.so/lyrasis/Add-LCP-support-to-Android-App-50c3d6cf123145d5be4d804b2cc929bd