Add LCP Support (#7)
* Initial LCP support

This adds the initial support for LCP in that it sets up the right
repositories and imports the correct LCP dependencies. It adds code
to retrieve an LcpService, if the LCP library is available, and
adds the needed format names and flags in order to ensure that
books are seen in feeds.

Affects: https://www.notion.so/lyrasis/Add-LCP-support-to-Android-App-50c3d6cf123145d5be4d804b2cc929bd

* Initial LCP fulfillment support

This adds the necessary support for fulfilling books using LCP.
Currently, the books cannot be decrypted as this requires the
registration of a ContentProtectionProvider.

Affects: https://www.notion.so/lyrasis/Add-LCP-support-to-Android-App-50c3d6cf123145d5be4d804b2cc929bd

* Complete LCP fulfillment

This completes LCP fulfillment support, including implementing the
"Automated Passphrase Retrieval" algorithm. Currently, the credentials
are not passed to R2, so books will fail to unlock.

Affects: https://www.notion.so/lyrasis/Add-LCP-support-to-Android-App-50c3d6cf123145d5be4d804b2cc929bd

* LCP decryption

This implements all of the code necessary to pass credentials to the
LCP decryption code.

Affects: https://www.notion.so/lyrasis/Add-LCP-support-to-Android-App-50c3d6cf123145d5be4d804b2cc929bd

* Regenerated the module list

Co-authored-by: Ray Lee <ray.lee@lyrasis.org>