Add the ability to specify a custom library registry
This adds a debugging interface and a bit of UI in the debug settings
menu that allows for setting a custom library registry server. The
setting does not persist after an app restart.

Affects: https://www.notion.so/lyrasis/Add-ability-in-Palace-Android-app-to-configure-library-registry-libraries-URL-7f88c7e8d0c84efe9b6fcc9906e4605a