Merge upstream changes (#18)
* Rewrite the announcements UI in a lifecycle-aware manner

- The announcements won't be all considered as acknowledged in case of configuration change.
- The announcements won't be all considered as acknowledged if only some of them have been checked by the patron

ProfileType.mostRecentAccount becomes non-nullable, and the profile database format is updated accordingly.
A profile already had to have at least one account, so now the most recent account is set to be this account when the profile is created.

* Fix an unused format string in the account deletion error message

* Use EULA specific to library instead of the global one in AccountDetailFragment.

* Refactor AccountDetailFragment and inadvertently fix the bug where the Login button shows 'Log in' while the patron is already logged in when resuming the Fragment

* Begin next development iteration

* Begin new changelog release (7.0.1)

* Use new R2 release

This switches to the latest R2 release to improve accessibility. It
conditionally enables continuous scrolling if the accessibility
service indicates that something equivalent to TalkBack is enabled.

Affects: https://jira.nypl.org/browse/SMA-197

* Switch DRM from Adept to AxisNow in OpenEbooks

* Refactor the CardCreatorService and a bit of the AccountDetailFragment

* Remove meaningless EULA checkbox from AccountDetailFragment

* ktlint

* Revert blanking of Barcode and PIN before logging out has completed

* Added extra check for displaying holds tab

* Cleaned up code check

* Make the account bookmark sync switch reflect reality

This adjusts the account bookmark sync switch so that it works based
off of an observable status value. When the switch is toggled, the
reader bookmark service tracks the progress of the operation and can
publish appropriate status values when asked. A small indefinite
progress indicator has been added to show that a long-running
operation is occurring when the switch is toggled, and the switch is
marked as disabled (non-interactable) when a sync setting change is
in progress. This prevents users from being able to queue up dozens
of switch changes by hammering the toggle repeatedly.

Affects: https://jira.nypl.org/browse/SMA-186

* Attempt to enable Open eBooks Firebase publishing

Affects: https://jira.nypl.org/browse/OE-163

* Set the Open eBooks version name

* Switch to the latest R2 release

This switches to the latest R2 release and updates the changelog.

* Fix a ktlint issue

Co-authored-by: Quentin Gliosca <quentin.gliosca@gmail.com>
Co-authored-by: Mark Raynsford <code@io7m.com>
Co-authored-by: qnga <32197639+qnga@users.noreply.github.com>
Co-authored-by: Maurice Edwards <me@mauricee.me>
Co-authored-by: Mark Raynsford <io7m@users.noreply.github.com>