Do not fail if position parsing fails
This upgrades to a version of the audio books API that will not fail
to parse positions that contain explicitly `null` titles. It also logs
an error and returns nothing instead of loudly raising an exception
if the saved position turns out to be malformed.

Affects: https://www.notion.so/lyrasis/Error-opening-app-after-exiting-from-audiobook-player-5694f5b2d40c4baea0d97782a3141cef