Gracefully handle expired certificates
This updates the boot code and DRM service to check for and reject
expired certificates. It adds a new boot event type that allows for
displaying a dialog if something important but not fatal occurs on
boot, and then publishes an event if the DRM fails.

Affects: https://www.notion.so/lyrasis/Android-Properly-handle-expired-Adobe-certificate-427ab218209d40d3856246cb220fdbca
Affects: https://www.notion.so/lyrasis/App-crashes-on-login-when-Adobe-cert-is-expired-e71eda890b034802ae6ceab83d363d1a