Stop using a boot event to show dialogs
The boot event stream is inherently lossy and racey, so it seems better
to show a dialog explicitly from the onStart() method of the main
fragment. The dialog is given one chance to display (and isn't subject
to race conditions), and isn't shown again until the app is restarted.

Affects: https://www.notion.so/lyrasis/Android-Properly-handle-expired-Adobe-certificate-427ab218209d40d3856246cb220fdbca
Affects: https://www.notion.so/lyrasis/App-crashes-on-login-when-Adobe-cert-is-expired-e71eda890b034802ae6ceab83d363d1a