Continue the neutral color scheme module
This aggressively removes the entirety of the old theme system
and replaces it with a neutral color scheme. This introduces a
new Toolbar subclass that implements the subset of behaviour
that we actually need from the toolbar.

Note that this will, currently, probably break the R2 and
audio book viewers as they have not yet been updated to use
the new theme information.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b