Neutral color scheme (#29)
* Introduce a neutral color scheme module

This introduces a neutral color scheme module and updates the tabs
to use it.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b

* Continue the neutral color scheme module

This aggressively removes the entirety of the old theme system
and replaces it with a neutral color scheme. This introduces a
new Toolbar subclass that implements the subset of behaviour
that we actually need from the toolbar.

Note that this will, currently, probably break the R2 and
audio book viewers as they have not yet been updated to use
the new theme information.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b

* Introduce Open Sans

This imports the Open Sans font and specifies it as the
application-wide global font.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b

* Adjust toolbar behaviour

This adjusts the toolbar behaviour to get the desired account selection
behaviour on different screens. It also fixes the incorrectly coloured
menu icons.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b

* Implement logos in feeds

This adds a section above feeds that show the logo and title of the
current library.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b

* Regenerate module list