Accounts -> Libraries terminology
This adjusts all occurrences in the UI of the word "account" and
changes it to "library". This also adjusts the toolbar to display
the tab label rather than the library name.

Affects: https://www.notion.so/lyrasis/Android-Update-Title-menu-Bar-lable-to-for-view-names-vs-library-names-and-catalog-breadcrumbs-cfdfbdb03614475984eef1b70a937f3a
Affects: https://www.notion.so/lyrasis/Android-Update-UI-label-for-Accounts-to-Libraries-90cfef17026044549c7b55c3b50451a2