Sync with upstream (#36)
* Rewrite the announcements UI in a lifecycle-aware manner

- The announcements won't be all considered as acknowledged in case of configuration change.
- The announcements won't be all considered as acknowledged if only some of them have been checked by the patron

ProfileType.mostRecentAccount becomes non-nullable, and the profile database format is updated accordingly.
A profile already had to have at least one account, so now the most recent account is set to be this account when the profile is created.

* Fix an unused format string in the account deletion error message

* Use EULA specific to library instead of the global one in AccountDetailFragment.

* Refactor AccountDetailFragment and inadvertently fix the bug where the Login button shows 'Log in' while the patron is already logged in when resuming the Fragment

* Begin next development iteration

* Begin new changelog release (7.0.1)

* Use new R2 release

This switches to the latest R2 release to improve accessibility. It
conditionally enables continuous scrolling if the accessibility
service indicates that something equivalent to TalkBack is enabled.

Affects: https://jira.nypl.org/browse/SMA-197

* Switch DRM from Adept to AxisNow in OpenEbooks

* Refactor the CardCreatorService and a bit of the AccountDetailFragment

* Remove meaningless EULA checkbox from AccountDetailFragment

* ktlint

* Revert blanking of Barcode and PIN before logging out has completed

* Added extra check for displaying holds tab

* Cleaned up code check

* Make the account bookmark sync switch reflect reality

This adjusts the account bookmark sync switch so that it works based
off of an observable status value. When the switch is toggled, the
reader bookmark service tracks the progress of the operation and can
publish appropriate status values when asked. A small indefinite
progress indicator has been added to show that a long-running
operation is occurring when the switch is toggled, and the switch is
marked as disabled (non-interactable) when a sync setting change is
in progress. This prevents users from being able to queue up dozens
of switch changes by hammering the toggle repeatedly.

Affects: https://jira.nypl.org/browse/SMA-186

* Attempt to enable Open eBooks Firebase publishing

Affects: https://jira.nypl.org/browse/OE-163

* Set the Open eBooks version name

* Switch to the latest R2 release

This switches to the latest R2 release and updates the changelog.

* Fix a ktlint issue

* Switch to the latest R2 release

This switches to the latest R2 release and updates the changelog.

Affects: https://jira.nypl.org/browse/SMA-212

* Temporarily disable Axis, temporarily enable Adobe

* Enable accounts UI

This switches the fallback server back to the production server,
and enables the UI to add and change accounts.

Affects: https://jira.nypl.org/browse/OE-140

* Restore OE accounts behaviour and colour scheme

* Schedule "revocation" for expired loans

This change hopefully fixes the issue where books are deleted from
the on-disk database, but the book registry is not updated properly
to indicate that loans are gone. This change simply removes the
database entry call, and schedules a revocation of the loan. The
revocation code _should_ be intelligent enough to correctly do any
and all cleanup work required.

Affects: https://jira.nypl.org/browse/SMA-205

* Use 'alternate' link for loan syncing

This undoes the previous commit and switches to a model where the book
registry for a given book is updated with up-to-the-minute-accurate
information on the server, using the "alternate" permalink for
each book. This should result in accurate book states when removing
books due to expired loans.

Affects: https://jira.nypl.org/browse/SMA-205

* Fix a crash related to stale account IDs

* Catch any possible exception raised when configuring Crashlytics.
* Fix broken account IDs when initializing profiles.

Affects: https://jira.nypl.org/browse/SMA-226

* Re-enable Axis DRM and set version to 2.1.0-SNAPSHOT for OE

* Add a toggle to enable/disable publisher CSS

Affects: https://jira.nypl.org/browse/SMA-229

* Disable book returns in OE

This adds a (deprecated) method to the catalog build configuration
service that allows for disabling book returns in applications.
This is only used by OE, and probably won't be used for all that long.

Affects: https://jira.nypl.org/browse/OE-149

* Filter unsupported local books

This adds the extra support required to filter out books that
were supported when they were fulfilled, but may be unsupported
now (due to a DRM system becoming unsupported).

Affects: https://jira.nypl.org/browse/OE-182

* Add test suite files missing from previous commit

* Check snapshot dependencies in non-snapshot builds

This implements a check in the build system that refuses to allow
non-SNAPSHOT builds of the app to build with -SNAPSHOT dependencies.
Essentially, this enforces reproducibility for release builds.

Fix: https://jira.nypl.org/browse/SIMPLY-3634

* Update changelog

* Implement an OE QA toggle

This implements a toggle that will add and remove an Open eBooks
QA account in the debug menu. Hardcoding the account information,
unfortunately, appears to be the only way to implement this as
there isn't a registry present in OE that seems to know about the OE
QA server.

Fix: https://jira.nypl.org/browse/OE-140

* Fix some lifecycle issues in Reader2Activity

* Delete book data for books with unsupported DRM systems

This updates the book database to be aware of book and DRM formats. If
an attempt is made to open a book that has an unsupported type of
DRM, the book data and the DRM information is deleted and the book is
reopened. The intention is that this will cause the book to appear
as having been loaned but not downloaded, and the user will then
re-download the book using the correct DRM system.  Bookmarks and
other information are preserved.

Affects: https://jira.nypl.org/browse/OE-192

* Adjust code for better debugging

* Correct code from previous commit

* Correct ktlint issue

* Upgrade the audiobook player

* Update changelog

* Add "Have you updated the changelog?"

* Updated dependency libraries

* Switch to Simplified-R2-Android 0.0.24-SNAPSHOT

* Update the changelog

* Update changelog

* Trim whitespace around usernames and passwords in AccountDetailFragment

* Update changelog

* Remove legacy R1 activities from manifests

* Add SMA-237 to changelog

* Revved snapshot version

* Update CI configuration

* Always produce AAB files

* Upload APK files to Firebase

* Turn off daily DRM builds

* Upgrade r2 libraries to 1.0.0-SNAPSHOT.

* Remove unneeded simplye and openebooks additions.

* Fix test compilation errors.

* Adjust a test to work for Palace.

We automatically delete the default account when a new account is created, but the test did not expect this. The test was adjusted to create two accounts, so that after one is deleted, there would still be a known account in the profile.

Co-authored-by: Quentin Gliosca <quentin.gliosca@gmail.com>
Co-authored-by: Mark Raynsford <code@io7m.com>
Co-authored-by: qnga <32197639+qnga@users.noreply.github.com>
Co-authored-by: Maurice Edwards <me@mauricee.me>
Co-authored-by: Mark Raynsford <io7m@users.noreply.github.com>
Co-authored-by: malcolmmcfly <mjw0133@gmail.com>