Hide "reload" menu item in catalog
This sets the "reload" menu item in the catalog to invisible. The
reason for making it invisible instead of deleting it is that deleting
the item would have meant deleting code that would otherwise have to
be reimplemented for "pull down to refresh" a few commits from now.

Affects: https://www.notion.so/lyrasis/Update-Android-UI-to-match-iOS-and-implement-new-icon-and-Brand-design-f504c6b7f5fe40c6a66c40a6e1c3438b