Implement an accounts graveyard
This changes the behaviour around deleting corrupted accounts. Now,
if an account is deleted because the app believes it to be corrupt,
it is instead moved into an "accounts-graveyard" directory stored next
to the normal "accounts" directory.  This will give us the opportunity
to grab corrupted accounts and see what caused them to break in the
first place.

Affects: https://www.notion.so/lyrasis/Error-code-51000-An-error-occurred-during-playback-on-Android-Audiobooks-db0d12b05e8f4d269011a776224b049a