Do not choke on malformed last-read locations
This adds a catch statement to avoid crashing the app if a last-read
location is unreadable. There are likely more instances of this kind
of behaviour in the app, but we have at least fixed this one.

Affects: https://www.notion.so/lyrasis/Error-code-51000-An-error-occurred-during-playback-on-Android-Audiobooks-db0d12b05e8f4d269011a776224b049a