Aggressively refactor the bookmark service
This refactors the implementation of the bookmark service to much
more aggressively save and restore bookmarks both locally and
to/from the server. Bookmarks are now regularly synced for all
enabled accounts, syncing exhaustively sends and receives all
bookmarks, and the following actions will trigger a sync:

  * Enabling syncing for an account.
  * Selecting a profile (essentially the same as opening the
    app in Palace builds).
  * Logging in to an account.
  * The app is left open for an hour. Repeats hourly.

The old functional policy implementation was removed due to
being overengineered and incomprehensible. The new implementation
is API-compatible, and all of the current unit tests pass.

Affects: https://www.notion.so/lyrasis/Palace-does-not-remember-place-in-book-or-audiobook-between-devices-on-Android-be7b4d5f3351465ca1f90ff480c6ad16