Default audio book time tracking to "on"
This sets the default for audio book time tracking to be
on instead of off. Note that, for installations that have
been used during development of the time tracking feature,
there might already be a setting of "off" saved to the
device, so this will require manually toggling the setting
back on again.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-328