Refactor bookmark service
This creates a new "create bookmark" operation in the bookmark service
that will first create a local bookmark, and then try to create an
equivalent remote bookmark. The operation can be configured to ignore
remote failures. This refactors the existing operations to return
either non-nullable Bookmarks, or fail the resulting Future values
with exceptions.

This also refactors the "delete bookmark" operation similarly. The
deletion occurs in reverse order; a remote bookmark is deleted first,
and then the corresponding local bookmark is deleted if the remote
bookmark deletion succeeds. This is also, currently, configured to
ignore remote failures at the current call sites.

These changes allow for having the bookmark service do the heavy
lifting, and simplifies the various viewers that use it.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-186