Handle AudioBook activity resources
This adjusts the AudioBookPlayerActivity to be more rigorous with
regards to resource closing and the UI thread. This implements a
simple stack of resources that are registered as they are created and
popped/closed on activity destruction. Additionally, all scheduled
UI thread operations are guarded by an isUISafe() method. This should
eliminate or at least reduce lifecycle-related crashes.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-656