Band-aid solution for bookmark deletion
This applies a band-aid to bookmark deletion. There are many cases
where the bookmark system might be handed the local copy of a bookmark
(without a server URI) instead of the server-side copy of the bookmark.
This band-aid solution searches for an equivalent bookmark and deletes
it remotely.

The real fix is to replace the server-side bookmark system with
something not designed by the W3C, and replace the client side code
entirely.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-1721