Accept 404 as a permanent error in time tracking
If a single entry in a time tracking response has a 404 or 401
status, the entry is considered to be permanently unusable and
will be deleted.

If the server response with an overall 404 status code, a fake
response will be synthesized with every entry in the request
transformed to an entry in the response with a 404 status code.
These entries will then be considered permanently failed and
deleted.

Fix: https://ebce-lyrasis.atlassian.net/browse/PP-2128