Stop waiting for bookmarks (PP-2932)
This eliminates calls to the bookmark service that wait for bookmarks
to be loaded. The bookmark service now publishes a read-only attribute
that consists of every bookmark in the system. The attribute is updated
as bookmarks are created and deleted (and synced to/from the server).

This should eliminate at least one ANR we're seeing when the app is
brought back into the foreground.